chore: Update npm dependencies and add isomorphic-dompurify for sanitizing HTML
