+++ apps/client/app/(dashboard)/gitcoin/[slug]/description.tsx
@@ -1,15 +1,15 @@
 'use client';
 
-import { renderToHTML } from 'app/(dashboard)/(_utils)/utils';
+import dynamic from 'next/dynamic';
+
+const DescriptionHtml = dynamic(() => import('@/components/gitcoin/description-html'), {
+  ssr: false,
+});
 
 export default function Description({ description }: { description: string }) {
   return (
     <div className="overflow-hidden gitcoindescriptionparent">
-      <div
-        dangerouslySetInnerHTML={{
-          __html: renderToHTML(description),
-        }}
-      />
+      <DescriptionHtml description={description} />
     </div>
   );
 }
