+++ apps/client/app/(dashboard)/gitcoin/[slug]/page.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { fetchApplicationById, fetchRoundForExplorer } from '@/lib/actions';
 import { Text } from '@mantine/core';
 import Description from './description';
