+++ apps/client/package.json
@@ -97,6 +97,7 @@
     "htmlparser2": "^9.0.0",
     "image-size": "^1.0.2",
     "input-otp": "^1.2.4",
+    "isomorphic-dompurify": "^2.14.0",
     "ldrs": "^1.0.1",
     "markdown-it": "^14.1.0",
     "marked": "^12.0.2",
